102-fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,12 +9,12 @@
 int main(void)
 {
 	int count;
-	int fib1 = 0, fib2 = 1, sum;
+	unsigned long fib1 = 0, fib2 = 1, sum;
 
 	for (count = 0; count < 50; count++)
 	{
 		sum = fib1 + fib2;
-		printf("%u", sum);
+		printf("%lu", sum);
 
 		fib1 = fib2;
 		fib2 = sum;
